feat(server): begin relocating server creation and export
